Limit concurrent app loading to prevent Binder flooding
Introduced a Semaphore with 4 permits in `AppRepository.loadApps` to limit the number of concurrent `mapToAppDetailed` calls. This prevents overwhelming system services (Binder flooding) when loading detailed information for a large number of installed applications.

- Added `MAX_CONCURRENCY = 4` constant.
- Used `semaphore.withPermit` inside the async loop.
- Added test case `verify concurrency limit` to `AppRepositoryTest` to verify the fix.
