+++ app/src/test/kotlin/com/github/keeganwitt/applist/AppRepositoryTest.kt
@@ -10,12 +10,17 @@ import com.github.keeganwitt.applist.services.UsageStatsService
 import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.toList
+import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.test.runTest
+import kotlinx.coroutines.withContext
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Before
 import org.junit.Test
+import java.util.concurrent.atomic.AtomicInteger
+import kotlin.math.max
 
 class AppRepositoryTest {
     private lateinit var packageService: PackageService
@@ -365,11 +370,11 @@ class AppRepositoryTest {
 
             val flow = repository.loadApps(
                 field = AppInfoField.VERSION,
-                showSystemApps = false, 
+                showSystemApps = false,
                 descending = false,
-                reload = false
+                reload = false,
             )
-            
+
             val result = flow.toList()
             val finalDocs = result.last()
 
@@ -378,6 +383,55 @@ class AppRepositoryTest {
             // Verify we got the basic info back
             assertEquals("Error App", finalDocs[0].name)
         }
+
+    @Test
+    fun `verify concurrency limit`() = runBlocking {
+        val appCount = 200
+        val apps = (1..appCount).map { i ->
+            ApplicationInfo().apply {
+                packageName = "com.app.$i"
+                flags = 0
+                enabled = true
+            }
+        }
+
+        every { packageService.getInstalledApplications(any<Long>()) } returns apps
+        every { packageService.loadLabel(any()) } returns "App"
+        every { packageService.getLaunchIntentForPackage(any()) } returns mockk()
+
+        // Concurrency tracker
+        val activeCoroutines = AtomicInteger(0)
+        val maxConcurrency = AtomicInteger(0)
+
+        // Mock a heavy operation
+        every { packageService.getPackageInfo(any()) } answers {
+            val current = activeCoroutines.incrementAndGet()
+            maxConcurrency.updateAndGet { prev -> max(prev, current) }
+
+            // Simulate blocking work
+            Thread.sleep(10)
+
+            activeCoroutines.decrementAndGet()
+            PackageInfo().apply {
+                versionName = "1.0"
+                firstInstallTime = 0
+                lastUpdateTime = 0
+            }
+        }
+
+        // Run on IO dispatcher to allow concurrency
+        withContext(Dispatchers.IO) {
+            repository.loadApps(
+                field = AppInfoField.VERSION,
+                showSystemApps = true,
+                descending = false,
+                reload = false
+            ).toList()
+        }
+
+        val maxObserved = maxConcurrency.get()
+        assertTrue("Max concurrency should be limited to 4, but was $maxObserved", maxObserved <= 4)
+    }
 
     private fun createApplicationInfo(
         packageName: String,
